Moving unique_schema, profiles_config_update, and cleanup to conftest.py
